// src/api/dtos/PontoEletronicoDTO.ts
export interface PontoEletronicoDTO {
  EscalaMotoristaId: string;
  Latitude: number;
  Longitude: number;
  CurrentTime: string;
  DataEntrada: string;
  DataSaida: string;
  Tipo: string;
  EmployeeId: string;
  LocalSaidaProgramadoId: string;
  Digital: string;
  Foto: string;
  CodigoFuncionario: string;
  CodigoBatePonto?: string;
}